# Using GL fork of gomemcache