**What this PR does**:

We see many times SG using lots of CPU fetching data from memcache and this fork has multiples opmizations to improve this scenario:



Ex: On the below flame graph we can see that sscanf is using lots of CPU:

![image](https://github.com/cortexproject/cortex/assets/4027760/f6880037-db33-42e1-abee-62477440c93d)

[This](https://github.com/grafana/gomemcache/commit/c3672f6c873893b96b313cffc77006c140792b32)  change improve this code path:

```
goos: linux
goarch: amd64
pkg: github.com/cortexproject/cortex/pkg/chunk/cache
cpu: Intel(R) Xeon(R) Platinum 8175M CPU @ 2.50GHz
        │   /tmp/old    │                /tmp/new                 │
        │    sec/op     │    sec/op     vs base                   │
SScanf    1676.0n ± ∞ ¹   710.2n ± ∞ ¹  -57.63% (p=0.016 n=4+5)
SScanF     1.674µ ± ∞ ¹
geomean    1.675µ         710.2n        -57.63%                 ²
¹ need >= 6 samples for confidence interval at level 0.95
² benchmark set differs from baseline; geomeans may not be comparable

        │  /tmp/old   │                /tmp/new                │
        │    B/op     │    B/op      vs base                   │
SScanf    94.00 ± ∞ ¹   16.00 ± ∞ ¹        ~ (p=0.079 n=4+5)
SScanF    94.00 ± ∞ ¹
geomean   94.00         16.00        -82.98%                 ²
¹ need >= 6 samples for confidence interval at level 0.95
² benchmark set differs from baseline; geomeans may not be comparable

        │  /tmp/old   │                /tmp/new                │
        │  allocs/op  │  allocs/op   vs base                   │
SScanf    5.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=0.079 n=4+5)
SScanF    5.000 ± ∞ ¹
geomean   5.000         1.000        -80.00%                 ²
¹ need >= 6 samples for confidence interval at level 0.95
```

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [NA] Tests updated
- [NA] Documentation added
- [NA] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`