@yeya24 2024-03-08
Let's close this until we find a way to upstream this change to Thanos.